==> ipynb/0_Overview of Result.ipynb <==
import pandas as pd
import warnings
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

compare = pd.concat([pd.read_csv('../result/comparing_HCD35_pos.csv', index_col=0),pd.read_csv('../result/comparing_HCD45_pos.csv', index_col=0),pd.read_csv('../result/comparing_HCD65_pos.csv', index_col=0),pd.read_csv('../result/comparing_HCD35_neg.csv', index_col=0),pd.read_csv('../result/comparing_HCD45_neg.csv', index_col=0),pd.read_csv('../result/comparing_HCD65_neg.csv', index_col=0),], axis=1)
compare.columns = ['HCD35_Positive','HCD45_Positive','HCD65_Positive','HCD35_Negative','HCD45_Negative','HCD65_Negative']
compare = round(compare*100, 2)
binning = pd.read_csv('../result/binning.csv', index_col=0)
k = pd.read_csv('../result/Feature_Selection_voting.csv', index_col=0)
w = k.rename(index={'Xgboost':'XGBoost'})
h = pd.read_csv('../result/Decomposition + HCD35_voting.csv', index_col=0)
p = h.rename(index={'Xgboost':'XGBoost'})
n = pd.read_csv('../result/Feature Selection + Decomposition_voting.csv', index_col=0)
v = n.rename(index={'Xgboost':'XGBoost'})
gs_pos = pd.read_csv('../result/voting_gs.csv', index_col=0)
gs_neg = pd.read_csv('../result/voting_gs_neg.csv', index_col=0)

### Binning

binning

### Comparing compound's num

compare

### Single model / Voting

pd.read_csv('../result/voting.csv', index_col=0).sort_values(['HCD35_Positive','HCD35_Negative'], ascending=False)

### Feature Selection & Decomposition / Voting

warnings.filterwarnings('ignore')
pd.concat([w, p, v], axis=1).sort_values(['Feature_Selection_Positive','Feature Selection + Decomposition_Positive'], ascending=False)

### GridSearch / Voting

warnings.filterwarnings('ignore')
pd.concat([gs_pos, gs_neg], axis=1).sort_values(['Feature_Selection_Positive','Feature Selection + Decomposition_Negative'], ascending=False)
