==> ipynb/1_Preprocess.ipynb <==
import numpy as np
import pandas as pd

# Load initial data
t = pd.read_csv('../data/itree_qexactive_data.csv').drop('25 HCD', axis=1)

# 65 HCD have na, so just dropna()
df = t.dropna()

# round off to two decimal places
df['Precursor m/z'] = round(df['Precursor m/z'], 2)

# check dataframe's shape
df.shape

# divide pos and neg
pos = df[df.Mode == 'positive'].reset_index(drop=True)
neg = df[df.Mode != 'positive'].reset_index(drop=True)

# drop_duplicates by InChIKey to get InChIKey list
pos.drop_duplicates('InChIKey').InChIKey.to_csv('../data/pos_key.csv')
neg.drop_duplicates('InChIKey').InChIKey.to_csv('../data/neg_key.csv')

# make class's list (just merge pos's class and neg's class
c = pd.concat([
    pd.read_csv('../data/pos_class.csv', index_col=None), 
    pd.read_csv('../data/neg_class.csv', index_col=None)]
).drop_duplicates('InChIKey').reset_index(drop=True)

# merge class and ms data by InChIKey
pos = pd.merge(pos, c, on='InChIKey')
neg = pd.merge(neg, c, on='InChIKey')

print(pos.shape)
print(neg.shape)

# drop_duplicates by precursor and InChIKey
pos = pos.drop_duplicates(subset=['Precursor m/z', 'InChIKey'])
neg = neg.drop_duplicates(subset=['Precursor m/z', 'InChIKey'])

print(pos.shape)
print(neg.shape)

pos.columns

# just choose 5 colomns to use as feature
pos[[
    '35 HCD','45 HCD','65 HCD','Kingdom', 
    'Superclass', 'Class','Subclass'
]].reset_index(drop=True).to_csv('../data/pos_for_ml.csv')
neg[[
    '35 HCD','45 HCD','65 HCD','Kingdom', 
    'Superclass', 'Class','Subclass'
]].reset_index(drop=True).to_csv('../data/neg_for_ml.csv')
