==> ipynb/2_Dataset.ipynb <==
import numpy as np
import pandas as pd
import pickle
import collections

# read csv
pos = pd.read_csv('../data/pos_for_ml.csv', index_col=0)
neg = pd.read_csv('../data/neg_for_ml.csv', index_col=0)

# choose top 4 subclass as target label
subclass = pos['Subclass'].value_counts()[:4].index
df_pos = pos[pos['Subclass'].isin(subclass)].reset_index(drop=True)

subclass = neg['Subclass'].value_counts()[:4].index
df_neg = neg[neg['Subclass'].isin(subclass)].reset_index(drop=True)

# change dict to dataframe as dataset
f = []
for p in df_pos['35 HCD']:
    binned_spectrum = collections.defaultdict(float)

    for ion in p.split(' '):
        mz, intensity = list(map(float, ion.split(':')))
        binned_spectrum[round(mz, 2)] += intensity
        
    # add dict to list 
    f.append(binned_spectrum)
    
t = pd.DataFrame(f).fillna(0)
t.columns = ['HCD35_'+str(i) for i in t.columns]
t = pd.concat([df_pos['Subclass'], t], axis=1)
t.to_pickle('../data/HCD35_pos.pickle')

# change dict to dataframe as dataset
f = []
for p in df_pos['45 HCD']:
    binned_spectrum = collections.defaultdict(float)

    for ion in p.split(' '):
        mz, intensity = list(map(float, ion.split(':')))
        binned_spectrum[round(mz, 2)] += intensity
        
    # add dict to list 
    f.append(binned_spectrum)
    
h = pd.DataFrame(f).fillna(0)
h.columns = ['HCD45_'+str(i) for i in h.columns]    
h = pd.concat([df_pos['Subclass'], h], axis=1)
h.to_pickle('../data/HCD45_pos.pickle')

# change dict to dataframe as dataset
f = []
for p in df_pos['65 HCD']:
    binned_spectrum = collections.defaultdict(float)

    for ion in p.split(' '):
        mz, intensity = list(map(float, ion.split(':')))
        binned_spectrum[round(mz, 2)] += intensity
        
    # add dict to list 
    f.append(binned_spectrum)
    
y = pd.DataFrame(f).fillna(0)
y.columns = ['HCD65_'+str(i) for i in y.columns]
y = pd.concat([df_pos['Subclass'], y], axis=1)
y.to_pickle('../data/HCD65_pos.pickle')

# change dict to dataframe as dataset
f = []
for p in df_neg['35 HCD']:
    binned_spectrum = collections.defaultdict(float)

    for ion in p.split(' '):
        mz, intensity = list(map(float, ion.split(':')))
        binned_spectrum[round(mz, 2)] += intensity
        
    # add dict to list 
    f.append(binned_spectrum)
    
t = pd.DataFrame(f).fillna(0)
t.columns = ['HCD35_'+str(i) for i in t.columns]    
t = pd.concat([df_neg['Subclass'], t], axis=1)
t.to_pickle('../data/HCD35_neg.pickle')

# change dict to dataframe as dataset
f = []
for p in df_neg['45 HCD']:
    binned_spectrum = collections.defaultdict(float)

    for ion in p.split(' '):
        mz, intensity = list(map(float, ion.split(':')))
        binned_spectrum[round(mz, 2)] += intensity
        
    # add dict to list 
    f.append(binned_spectrum)
    
h = pd.DataFrame(f).fillna(0)
h.columns = ['HCD45_'+str(i) for i in h.columns]    
h = pd.concat([df_neg['Subclass'], h], axis=1)
h.to_pickle('../data/HCD45_neg.pickle')

# change dict to dataframe as dataset
f = []
for p in df_neg['65 HCD']:
    binned_spectrum = collections.defaultdict(float)

    for ion in p.split(' '):
        mz, intensity = list(map(float, ion.split(':')))
        binned_spectrum[round(mz, 2)] += intensity
        
    # add dict to list 
    f.append(binned_spectrum)
    
y = pd.DataFrame(f).fillna(0)
y.columns = ['HCD65_'+str(i) for i in y.columns]    
y = pd.concat([df_neg['Subclass'], y], axis=1)
y.to_pickle('../data/HCD65_neg.pickle')
