==> ipynb/4_LGBM.ipynb <==
import pickle
import pandas as pd
import numpy as np
import time

import xgboost as xgb
import lightgbm as lgb

from sklearn.ensemble import RandomForestClassifier as rf
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn.model_selection import GridSearchCV
from sklearn import preprocessing

# load data
with open('../data/HCD35_pos.pickle', mode='rb') as fp:
    df = pickle.load(fp)

# divide objective and target
objective = df.Subclass
le = preprocessing.LabelEncoder()
objective = le.fit_transform(objective)

features = df.drop('Subclass', axis=1)

# train test split
random_state=np.random.seed(42)
X_train, X_test, y_train, y_test = train_test_split(
    features, 
    objective,
    test_size=0.2
)

# define and fit
start = time.time()
gbm = lgb.LGBMClassifier(
    objective='multiclass',
    device = 'gpu',
    gpu_device_id=0,
    n_jobs=-1 
)

gbm.fit(
    X_train, 
    y_train,
    eval_set=[(X_test, y_test)],
    early_stopping_rounds=5
)
elapsed_time = time.time() - start

# save model to sav
pickle.dump(gbm, open('../model/lgbm_HCD35_pos.sav', 'wb'))

# add result to list 
o = {}
d = {}
o['HCD35_Positive'] = [round(gbm.score(X_test, y_test)*100, 2)]
d['HCD35_Positive'] = [elapsed_time]

# load data
with open('../data/HCD45_pos.pickle', mode='rb') as fp:
    df = pickle.load(fp)

# divide objective and target
objective = df.Subclass
le = preprocessing.LabelEncoder()
objective = le.fit_transform(objective)

features = df.drop('Subclass', axis=1)

# train test split
random_state=np.random.seed(42)
X_train, X_test, y_train, y_test = train_test_split(
    features, 
    objective,
    test_size=0.2
)

# define and fit
start = time.time()
gbm = lgb.LGBMClassifier(
    objective='multiclass',
    device = 'gpu',
    gpu_device_id=0,
    n_jobs=-1 
)

gbm.fit(
    X_train, 
    y_train,
    eval_set=[(X_test, y_test)],
    early_stopping_rounds=5
)
elapsed_time = time.time() - start

# save model to sav
pickle.dump(gbm, open('../model/lgbm_HCD45_pos.sav', 'wb'))

gbm.score(X_test, y_test)

# add result to list 
o['HCD45_Positive'] = [round(gbm.score(X_test, y_test)*100, 2)]
d['HCD45_Positive'] = [elapsed_time]

# load data
with open('../data/HCD65_pos.pickle', mode='rb') as fp:
    df = pickle.load(fp)

# divide objective and target
objective = df.Subclass
le = preprocessing.LabelEncoder()
objective = le.fit_transform(objective)

features = df.drop('Subclass', axis=1)

# train test split
random_state=np.random.seed(42)
X_train, X_test, y_train, y_test = train_test_split(
    features, 
    objective,
    test_size=0.2
)

# define and fit
start = time.time()
gbm = lgb.LGBMClassifier(
    objective='multiclass',
    device = 'gpu',
    gpu_device_id=0,
    n_jobs=-1 
)

gbm.fit(
    X_train, 
    y_train,
    eval_set=[(X_test, y_test)],
    early_stopping_rounds=5
)
elapsed_time = time.time() - start

# save model to sav
pickle.dump(gbm, open('../model/lgbm_HCD65_pos.sav', 'wb'))
gbm.score(X_test, y_test)

# add result to list 
o['HCD65_Positive'] = [round(gbm.score(X_test, y_test)*100, 2)]
d['HCD65_Positive'] = [elapsed_time]

# load data 
with open('../data/HCD35_neg.pickle', mode='rb') as fp:
    df = pickle.load(fp)

# divide objective and target
objective = df.Subclass
le = preprocessing.LabelEncoder()
objective = le.fit_transform(objective)

features = df.drop('Subclass', axis=1)

# train test split
random_state=np.random.seed(42)
X_train, X_test, y_train, y_test = train_test_split(
    features, 
    objective,
    test_size=0.2
)

# define and fit
start = time.time()
gbm = lgb.LGBMClassifier(
    objective='multiclass',
    device = 'gpu',
    gpu_device_id=0,
    n_jobs=-1 
)

gbm.fit(
    X_train, 
    y_train,
    eval_set=[(X_test, y_test)],
    early_stopping_rounds=5
)
elapsed_time = time.time() - start
# save model to sav
pickle.dump(gbm, open('../model/lgbm_HCD35_neg.sav', 'wb'))
gbm.score(X_test, y_test)

# result append to list
o['HCD35_Negative'] = [round(gbm.score(X_test, y_test)*100, 2)]
d['HCD35_Negative'] = [elapsed_time]

# load data
with open('../data/HCD45_neg.pickle', mode='rb') as fp:
    df = pickle.load(fp)

# divide objective and target
objective = df.Subclass
le = preprocessing.LabelEncoder()
objective = le.fit_transform(objective)

features = df.drop('Subclass', axis=1)

# train test split
random_state=np.random.seed(42)
X_train, X_test, y_train, y_test = train_test_split(
    features, 
    objective,
    test_size=0.2
)

# define and fit
start = time.time()
gbm = lgb.LGBMClassifier(
    objective='multiclass',
    device = 'gpu',
    gpu_device_id=0,
    n_jobs=-1 
)

gbm.fit(
    X_train, 
    y_train,
    eval_set=[(X_test, y_test)],
    early_stopping_rounds=5
)
elapsed_time = time.time() - start

# save model to sav
pickle.dump(gbm, open('../model/lgbm_HCD45_neg.sav', 'wb'))
gbm.score(X_test, y_test)

# result append to list
o['HCD45_Negative'] = [round(gbm.score(X_test, y_test)*100, 2)]
d['HCD45_Negative'] = [elapsed_time]

# load data
with open('../data/HCD65_neg.pickle', mode='rb') as fp:
    df = pickle.load(fp)

# divide objective and target
objective = df.Subclass
le = preprocessing.LabelEncoder()
objective = le.fit_transform(objective)

features = df.drop('Subclass', axis=1)

# train test split
random_state=np.random.seed(42)
X_train, X_test, y_train, y_test = train_test_split(
    features, 
    objective,
    test_size=0.2
)

# define and fit
start = time.time()
gbm = lgb.LGBMClassifier(
    objective='multiclass',
    device = 'gpu',
    gpu_device_id=0,
    n_jobs=-1 
)

gbm.fit(
    X_train, 
    y_train,
    eval_set=[(X_test, y_test)],
    early_stopping_rounds=5
)
elapsed_time = time.time() - start

# save model to sav
pickle.dump(gbm, open('../model/lgbm_HCD65_neg.sav', 'wb'))
gbm.score(X_test, y_test)

# result append to list
o['HCD65_Negative'] = [round(gbm.score(X_test, y_test)*100, 2)]
d['HCD65_Negative'] = [elapsed_time]

# change result list to Dataframe
g = pd.concat([pd.DataFrame(o).T, pd.DataFrame(d).T], axis=1)
g.columns = ['Accuracy', 'Time']

# dataframe to csv
g.to_csv('../result/LightGBM.csv')
